# web/frontend-v2/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Z-Image Generator",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/qwen-image/edit-layer": {
      "post": {
        "summary": "Qwen Image Edit Layer",
        "description": "Edit a decomposed layer using text instructions.\n\nUses the Qwen-Image-Edit-2511 model to modify a layer based on natural language\ninstructions. The edit model is loaded lazily on first use.\n\nReturns the edited RGBA layer as a PNG image.",
        "operationId": "qwen_image_edit_layer_api_qwen_image_edit_layer_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QwenImageEditLayerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/qwen-image/edit-multi": {
      "post": {
        "summary": "Qwen Image Edit Multi",
        "description": "Combine multiple images using Qwen-Image-Edit-2511.\n\nNew capability in Edit-2511 for multi-person consistency and creative\nimage merging. Supports combining 2+ images into a single coherent output.\n\nReturns the combined output as a PNG image.",
        "operationId": "qwen_image_edit_multi_api_qwen_image_edit_multi_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QwenImageEditMultiRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/qwen-image-2512/generate": {
      "post": {
        "summary": "Qwen Image 2512 Generate",
        "description": "Generate an image using Qwen-Image T2I (pure text-to-image).\n\nUses unified config: --model-type qwenimage-t2i --qwen-image-model-path\nVariant-aware defaults: T2I uses 40 steps, 1024 resolution, fp8 quantization.",
        "operationId": "qwen_image_2512_generate_api_qwen_image_2512_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QwenImage2512GenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QwenImage2512GenerateRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "negative_prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Negative Prompt"
          },
          "width": {
            "type": "integer",
            "title": "Width",
            "default": 1024
          },
          "height": {
            "type": "integer",
            "title": "Height",
            "default": 1024
          },
          "steps": {
            "type": "integer",
            "title": "Steps",
            "default": 40
          },
          "cfg_scale": {
            "type": "number",
            "title": "Cfg Scale",
            "default": 4.0
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          },
          "max_sequence_length": {
            "type": "integer",
            "title": "Max Sequence Length",
            "default": 512
          }
        },
        "type": "object",
        "required": [
          "prompt"
        ],
        "title": "QwenImage2512GenerateRequest",
        "description": "Request for Qwen-Image T2I text-to-image generation."
      },
      "QwenImageEditLayerRequest": {
        "properties": {
          "layer_image": {
            "type": "string",
            "title": "Layer Image"
          },
          "instruction": {
            "type": "string",
            "title": "Instruction"
          },
          "steps": {
            "type": "integer",
            "title": "Steps",
            "default": 40
          },
          "cfg_scale": {
            "type": "number",
            "title": "Cfg Scale",
            "default": 4.0
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          }
        },
        "type": "object",
        "required": [
          "layer_image",
          "instruction"
        ],
        "title": "QwenImageEditLayerRequest",
        "description": "Request for Qwen-Image layer editing (single image)."
      },
      "QwenImageEditMultiRequest": {
        "properties": {
          "images": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Images"
          },
          "instruction": {
            "type": "string",
            "title": "Instruction"
          },
          "steps": {
            "type": "integer",
            "title": "Steps",
            "default": 40
          },
          "cfg_scale": {
            "type": "number",
            "title": "Cfg Scale",
            "default": 4.0
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          }
        },
        "type": "object",
        "required": [
          "images",
          "instruction"
        ],
        "title": "QwenImageEditMultiRequest",
        "description": "Request for Qwen-Image multi-image editing (2511 feature)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}