# web/frontend-v2/src/api/types.ts
/**
 * API Types -- generated from OpenAPI spec + frontend-only types
 *
 * Backend types are generated by @hey-api/openapi-ts from the FastAPI
 * OpenAPI schema. Re-exported here with aliases for backward compat.
 *
 * Types that are kept hand-written are marked with [HAND] and a reason.
 * All others re-export from generated code so they stay in sync with
 * the backend automatically.
 */

// ---------------------------------------------------------------------------
// Re-exports from generated types (backend-derived, good fit)
// ---------------------------------------------------------------------------

export type {
  // LoRA types -- generated shapes match frontend usage exactly
  LoRaInfo as LoRAInfo,
  LoRaFileInfo as LoRAFile,
  LoRaListResponse as LoRAListResponse,

  // Simple response types
  ImageGenerationResult,
  ClearCacheResponse,

  // Config
  ConfigTag,
  ConfigWarning,
} from '../generated';

// ---------------------------------------------------------------------------
// [HAND] Types kept hand-written (generated versions too loose or missing fields)
// ---------------------------------------------------------------------------

// Param type -- determines which control renders
export type ParamType = 'textarea' | 'slider' | 'number' | 'checkbox' | 'select' | 'image' | 'color' | 'lora_list';

// Output type -- affects result display
export type OutputType = 'image' | 'video' | 'layers';

// Param group -- for progressive disclosure
export type GroupType = 'basic' | 'advanced' | 'expert' | 'scheduler' | 'optimization' | 'enhancement';

/**
 * [HAND] ParamSchema -- generated version uses `| null` for optional fields
 * but frontend expects `| undefined`. Also uses string literal unions for
 * type/group that generated version widens to `string`.
 */
export interface ParamSchema {
  id: string;
  type: ParamType;
  label: string;
  default?: unknown;
  min?: number;
  max?: number;
  step?: number;
  options?: string[];
  options_endpoint?: string;
  group: GroupType;
  tooltip?: string;
  conditional?: Record<string, unknown>;
  placeholder?: string;
  rows?: number;
  required?: boolean;
  max_count?: number;
  scale_min?: number;
  scale_max?: number;
  dependent_defaults?: Record<string, Record<string, unknown>>;
}

/**
 * [HAND] PipelineSchema -- generated uses camelCase (supportsStreaming) but
 * backend sends snake_case in the dict serialization (to_dict() method).
 * The backend returns raw dicts, not CamelModel instances, so snake_case
 * is what the frontend actually receives.
 */
export interface PipelineSchema {
  id: string;
  name: string;
  description: string;
  output_type: OutputType;
  color: string;
  icon?: string;
  params: ParamSchema[];
  supports_history: boolean;
  supports_img2img: boolean;
  supports_reference_images: boolean;
  supports_streaming: boolean;
  endpoint: string;
  category: string;
}

/**
 * [HAND] PipelinesResponse -- generated version types pipelines as
 * `{ [key: string]: unknown }` which loses all type info. We need
 * `Record<string, PipelineSchema>` for type-safe pipeline access.
 */
export interface PipelinesResponse {
  pipelines: Record<string, PipelineSchema>;
  defaults: Record<string, unknown>;
  loaded_pipeline: string | null;
}

/**
 * [HAND] PresetsResponse -- generated PresetListResponse types presets
 * as `Array<{ [key: string]: unknown }>`, losing GenerationPreset shape.
 */
export interface PresetsResponse {
  presets: GenerationPreset[];
  defaultPreset: string;
}

/**
 * [HAND] VRAMStatus -- generated version makes all fields optional
 * (because Pydantic defaults), but they're always present in practice.
 */
export interface VRAMStatus {
  usedMb: number;
  totalMb: number;
  freeMb: number;
  utilizationPercent: number;
  breakdown: {
    label: string;
    sizeMb: number;
    color: string;
  }[];
}

/**
 * [HAND] ModelStatusResponse -- generated version missing frontend-only
 * `error` field, and `pipelineId` is required in generated but we create
 * partial objects like `{ status: 'loading' }` in the store.
 */
export type ModelStatus = 'loaded' | 'unloaded' | 'loading' | 'error';

export interface ModelStatusResponse {
  pipelineId?: string;
  status: ModelStatus | string;
  components?: { name: string; sizeMb: number; device: string }[];
  totalVramMb?: number;
  vramMb?: number;
  modelVariant?: string | null;
  displayName?: string | null;
  loras?: { name: string; scale: number; layersUpdated: number }[];
  loraSummary?: string | null;
  configTags?: import('../generated').ConfigTag[];
  configWarnings?: import('../generated').ConfigWarning[];
  error?: string;
}

/**
 * [HAND] GenerationContext -- generated version makes all fields optional.
 * Frontend expects required fields for the status bar display.
 */
export interface GenerationContext {
  uptimeSeconds: number | null;
  profile: string;
  activePipeline: string | null;
  pipelineDisplayName: string | null;
  modelVariant: string | null;
  loras: import('../generated').LoRaInfo[];
  loraSummary: string | null;
  quantization: Record<string, string>;
  compileEnabled: boolean;
  compileMode: string | null;
  blockOffload: boolean;
  vramUsedGb: number | null;
  vramTotalGb: number | null;
  vramPercent: number | null;
  pendingRestartFields: string[];
  sessionModifiedFields: string[];
  fmttCached: boolean;
  historyCount: number;
}

/**
 * [HAND] GenerationPreset -- generated PresetDetailResponse makes `params`
 * optional (`params?: { [key: string]: unknown }`), but presets always
 * include params.
 */
export interface GenerationPreset {
  name: string;
  description: string;
  category: string;
  pipelines: string[];
  variant: string | null;
  params: Record<string, unknown>;
}

// ---------------------------------------------------------------------------
// Frontend-only types (no backend equivalent)
// ---------------------------------------------------------------------------

/**
 * Generation progress update (SSE event shape -- not in OpenAPI since SSE is untyped)
 */
export interface GenerationProgress {
  step: number;
  total: number;
  percent?: number;
  message?: string;
}

/**
 * Generation result -- frontend-enriched version of ImageGenerationResult.
 * Adds params, durationMs, timestamp fields for history tracking.
 */
export interface GenerationResult {
  id: string;
  pipelineId: string;
  outputType: OutputType;
  urls: string[];
  thumbnailUrl?: string;
  params: Record<string, unknown>;
  seed: number;
  durationMs: number;
  timestamp: number;
}

/**
 * Generation error (SSE error shape)
 */
export interface GenerationError {
  message: string;
  code?: string;
  details?: unknown;
  recoverable: boolean;
}

/**
 * History item for display (frontend IndexedDB shape)
 */
export interface HistoryItem {
  id: string;
  pipelineId: string;
  pipelineName: string;
  pipelineColor: string;
  thumbnailUrl: string;
  fullImageUrl?: string;
  prompt: string;
  shortPrompt: string;
  keyParams: string;
  timestamp: number;
  params: Record<string, unknown>;
  result: GenerationResult;
}

/**
 * Form values -- generic key-value for any pipeline
 */
export type FormValues = Record<string, unknown>;

/**
 * Validation error (frontend validation)
 */
export interface ValidationError {
  paramId: string;
  message: string;
}